pcm-pcie: fix compile error and warnings on Windows

+++ pcm-pcie.cpp
@@ -235,7 +235,7 @@ int main(int argc, char * argv[])
     cerr << "Update every " << delay << " seconds\n";
 
     unique_ptr<IPlatform> platform(IPlatform::getPlatform(m, csv, print_bandwidth,
-                                    print_additional_info, delay));
+                                    print_additional_info, (uint)delay)); // FIXME: do we support only integer delay?
 
     if (!platform)
     {
